feat: Simplex and large scale W4 WX rework

+++ Application/AstuteApplication.cpp
@@ -1,3 +1,5 @@
+#include "pch/ApplicationPCH.hpp"
+
 #include "AstuteApplication.hpp"
 
 #include <imgui.h>
@@ -37,7 +39,7 @@ AstuteApplication::AstuteApplication(Application::Configuration config)
 auto
 AstuteApplication::update(f64 ts) -> void
 {
-  camera->on_update(ts);
+  camera->on_update(static_cast<f32>(ts));
 
   switch (scene_state) {
     using enum AstuteApplication::SceneState;
@@ -104,15 +106,15 @@ AstuteApplication::interface() -> void
       light_environment.is_perspective ? "Perspective" : "Ortho";
     const std::string inverse_label =
       light_environment.is_perspective ? "Ortho" : "Perspective";
-    UI::coloured_text({ 0.1, 0.9, 0.6, 1.0 }, "Current chosen: {}", label);
+    UI::coloured_text({ 0.1F, 0.9F, 0.6F, 1.0F }, "Current chosen: {}", label);
     ImGui::Checkbox(inverse_label.c_str(), &light_environment.is_perspective);
     if (!light_environment.is_perspective) {
-      static f32 left{ -10 };
-      static f32 right{ 10 };
-      static f32 bottom{ -10 };
-      static f32 top{ 10 };
-      static f32 near{ 3 };
-      static f32 far{ 90 };
+      static f32 left{ -10.0F };
+      static f32 right{ 10.0F };
+      static f32 bottom{ -10.0F };
+      static f32 top{ 10.0F };
+      static f32 near{ 3.0F };
+      static f32 far{ 90.0F };
       ImGui::InputFloat("Left", &left);
       ImGui::InputFloat("Right", &right);
       ImGui::InputFloat("Bottom", &bottom);
@@ -122,10 +124,10 @@ AstuteApplication::interface() -> void
       auto projection = glm::ortho(left, right, bottom, top, near, far);
       light_environment.shadow_projection = projection;
     } else {
-      static f32 fov{ 45 };
-      static f32 aspect{ 1 };
-      static f32 near{ 0.1 };
-      static f32 far{ 100 };
+      static f32 fov{ 45.0F };
+      static f32 aspect{ 1.0F };
+      static f32 near{ 0.1F };
+      static f32 far{ 100.0F };
       ImGui::InputFloat("FOV", &fov);
       ImGui::InputFloat("Aspect", &aspect);
       ImGui::InputFloat("Near", &near);
